Dissolve `getTokenAndBalanceForSponsorship`
All its pieces are now in their rightful places.
