--- src/getters/getTokenAndBalanceForSponsorship.ts
@@ -1,17 +1,0 @@
-import { getCustomTokenBalance } from '~/marketplace/utils/web3'
-import { defaultChainConfig } from './getChainConfig'
-import getCoreConfig from './getCoreConfig'
-
-export type TokenAndBalanceForSponsorship = {
-    balance: string
-    tokenSymbol: string
-    tokenDecimals: number
-}
-
-export async function getBalanceForSponsorship(wallet: string) {
-    return getCustomTokenBalance(
-        defaultChainConfig.contracts[getCoreConfig().sponsorshipPaymentToken],
-        wallet,
-        defaultChainConfig.id,
-    )
-}
